generate README in the same step as the spark configuration
